+++ src/flat-configs/custom/custom-rules/eslint.ts
@@ -54,7 +54,6 @@ export const eslintRules: FlatConfigRules = {
   'no-sequences': 'error',
   'no-template-curly-in-string': 'error',
   'no-throw-literal': 'error',
-  'no-undef': 'off',
   'no-undef-init': 'error',
   'no-unmodified-loop-condition': 'error',
   'no-unneeded-ternary': ['error', { defaultAssignment: false }],