+++ src/flat-configs/custom/custom-rules/index.ts
@@ -4,6 +4,7 @@ import { eslintCommentsRules } from './eslint-comments.js'
 import { eslintRules } from './eslint.js'
 import { importRules } from './import.js'
 import { nRules } from './n.js'
+import { nonCompatibleRules } from './non-compatible.js'
 import { perfectionistRules } from './perfectionist.js'
 import { promiseRules } from './promise.js'
 import { reactRules } from './react.js'
@@ -59,6 +60,7 @@ export function getCustomRulesConfigs() {
   if (options.perfectionist) {
     Object.assign(customRules, perfectionistRules)
   }
+  Object.assign(customRules, nonCompatibleRules)
 
   const customRulesConfigs: FlatConfigs = [{ name: 'custom-rules', rules: customRules }]
   return customRulesConfigs