+++ src/flat-configs/custom/typescript.ts
@@ -1,0 +1,15 @@
+import { tsGlob } from '../../lib/constants.js'
+import type { FlatConfigs } from '../../types/eslint.js'
+
+export function getTypeScriptConfigs() {
+  const typeScriptConfigs: FlatConfigs = [
+    {
+      files: [tsGlob],
+      name: 'typescript',
+      rules: {
+        'no-undef': 'off',
+      },
+    },
+  ]
+  return typeScriptConfigs
+}